+++ backend/src/main/java/groom/backend/interfaces/coupon/dto/request/CouponCreateRequest.java
@@ -1,7 +1,7 @@
 package groom.backend.interfaces.coupon.dto.request;
 
-import groom.backend.domain.coupon.entity.Coupon;
-import groom.backend.domain.coupon.enums.CouponType;
+import groom.backend.domain.coupon.model.entity.Coupon;
+import groom.backend.domain.coupon.model.enums.CouponType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
