+++ backend/src/main/java/groom/backend/interfaces/coupon/dto/response/CouponResponse.java
@@ -1,9 +1,8 @@
 package groom.backend.interfaces.coupon.dto.response;
 
-import groom.backend.domain.coupon.entity.Coupon;
-import groom.backend.domain.coupon.enums.CouponType;
+import groom.backend.domain.coupon.model.entity.Coupon;
+import groom.backend.domain.coupon.model.enums.CouponType;
 import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
 import java.time.LocalDate;
