+++ backend/src/test/java/groom/backend/application/coupon/CouponIssueServiceIntegrationTest.java
@@ -5,9 +5,9 @@ import groom.backend.common.exception.ErrorCode;
 import groom.backend.domain.auth.entity.User;
 import groom.backend.domain.auth.enums.Grade;
 import groom.backend.domain.auth.enums.Role;
-import groom.backend.domain.coupon.entity.Coupon;
-import groom.backend.domain.coupon.entity.CouponIssue;
-import groom.backend.domain.coupon.enums.CouponType;
+import groom.backend.domain.coupon.model.entity.Coupon;
+import groom.backend.domain.coupon.model.entity.CouponIssue;
+import groom.backend.domain.coupon.model.enums.CouponType;
 import groom.backend.domain.coupon.repository.CouponIssueRepository;
 import groom.backend.domain.coupon.repository.CouponRepository;
 import groom.backend.interfaces.auth.persistence.SpringDataUserRepository;
